feat: 在插件加载时获取节假日信息
